chore: modularize field validators
